refactor: rename 'kiosk' to 'kiosk_utils'
