+++ lib/utils/kiosk_utils.dart
